+++ packages/manager/apps/container/src/container/legacy/server-sidebar/universe/DedicatedSidebar.tsx
@@ -10,6 +10,9 @@ import dedicatedShopConfig from '../order/shop-config/dedicated';
 import OrderTrigger from '../order/OrderTrigger';
 import { ShopItem } from '../order/OrderPopupContent';
 import  getIcon  from './GetIcon';
+import { OsdsIcon } from '@ovhcloud/ods-components/react';
+import { ODS_ICON_NAME, ODS_ICON_SIZE } from '@ovhcloud/ods-components';
+import { ODS_THEME_COLOR_INTENT } from '@ovhcloud/ods-common-theming';
 
 export const features = [
   'dedicated-server',
@@ -46,6 +49,7 @@ export const features = [
   'dedicated-server:nutanixOrder',
   'carbon-calculator',
   'network-security',
+  'key-management-service'
 ];
 
 export default function DedicatedSidebar() {
@@ -280,6 +284,7 @@ export default function DedicatedSidebar() {
             id: 'ip-loadbalancer',
             label: t('sidebar_pci_load_balancer'),
             icon: getIcon('ovh-font ovh-font-iplb'),
+            href: navigation.getURL('dedicated', '#/iplb'),
             routeMatcher: new RegExp('^(/network)?/iplb'),
             async loader() {
               const iplb = await loadServices('/ipLoadbalancing');
@@ -295,6 +300,7 @@ export default function DedicatedSidebar() {
             id: 'dedicated-vrack',
             label: t('sidebar_vrack'),
             icon: getIcon('ovh-font ovh-font-vRack'),
+            href: navigation.getURL('dedicated', '#/vrack'),
             routeMatcher: new RegExp('^/vrack'),
             async loader() {
               return loadServices('/vrack');
@@ -305,36 +311,22 @@ export default function DedicatedSidebar() {
             badge: 'beta',
             label: t('sidebar_vrack_services'),
             icon: getIcon('oui-icon oui-icon-vRack-services_concept'),
+            href: navigation.getURL('vrack-services', '/'),
             pathMatcher: new RegExp('^/vrack-services'),
             async loader() {
               const appId = 'vrack-services';
-              const items = await loadServices('/vrackServices/resource', undefined, appId);
-
-              return [
-                {
-                  id: 'vrack_services-all',
-                  label: t('sidebar_all_vrack_services'),
-                  href: navigation.getURL(appId, '#/'),
-                  ignoreSearch: true,
-                },
-                ...items
-              ];
+              return loadServices('/vrackServices/resource', undefined, appId);
             },
           },
           feature['cloud-connect'] && {
             id: 'dedicated-ovhcloudconnect',
             label: t('sidebar_cloud_connect'),
             icon: getIcon('oui-icon oui-icon-line-communicating_concept'),
+            href: navigation.getURL('dedicated', '#/cloud-connect'),
             routeMatcher: new RegExp('^/cloud-connect'),
             async loader() {
               const services = await loadServices('/ovhCloudConnect');
               return [
-                {
-                  id: 'cloud-connect-all',
-                  label: t('sidebar_service_all'),
-                  href: navigation.getURL('dedicated', '#/cloud-connect'),
-                  ignoreSearch: true,
-                },
                 ...services.map((service) => ({
                   ...service,
                   href: navigation.getURL(
@@ -394,6 +386,26 @@ export default function DedicatedSidebar() {
         icon: getIcon('ovh-font ovh-font-certificate'),
         href: navigation.getURL('dedicated', '#/license'),
         routeMatcher: new RegExp('/license'),
+      });
+    }
+
+    if (feature['key-management-service']) {
+      menu.push({
+        id: 'identity-security-operations',
+        label: t('sidebar_identity_security_operations'),
+        icon: <OsdsIcon name={ODS_ICON_NAME.CLOUD_EYE_CONCEPT} size={ODS_ICON_SIZE.xxs} color={ODS_THEME_COLOR_INTENT.text}/>,
+        badge: 'new',
+        pathMatcher: new RegExp('^/key-management-service'),
+        subItems: [
+          {
+            id: 'key-management-service',
+            label: t('sidebar_key-management-service'),
+            href: navigation.getURL('key-management-service', '/'),
+            badge: 'beta',
+            pathMatcher: new RegExp('^/key-management-service'),
+            icon: <OsdsIcon name={ODS_ICON_NAME.KEY_CONCEPT} size={ODS_ICON_SIZE.xxs} color={ODS_THEME_COLOR_INTENT.text}/>
+          },
+        ],
       });
     }