+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/navigation-tree/services/publicCloud.ts
@@ -1,12 +1,11 @@
 import { Node, NodeTag } from '../node';
+import illustration from '@/assets/images/sidebar/public-cloud.png';
 
 const pciNode: Node = {
   id: 'pci',
   translation: 'sidebar_pci',
-  routing: {
-    application: 'public-cloud',
-    hash: '#/pci/projects/{projectId}',
-  },
+  shortTranslation: 'sidebar_pci_short',
+  illustration,
   features: ['public-cloud'],
   forceVisibility: true,
   children: [
@@ -277,7 +276,8 @@ const pciNode: Node = {
           serviceType: 'CLOUD_PROJECT_DATABASE',
           routing: {
             application: 'public-cloud',
-            hash: '#/pci/projects/{projectId}/storages/databases-analytics/databases',
+            hash:
+              '#/pci/projects/{projectId}/storages/databases-analytics/databases',
           },
           features: ['databases'],
           forceVisibility: true,
@@ -288,7 +288,8 @@ const pciNode: Node = {
           serviceType: 'CLOUD_PROJECT_DATABASE',
           routing: {
             application: 'public-cloud',
-            hash: '#/pci/projects/{projectId}/storages/databases-analytics/data-streaming',
+            hash:
+              '#/pci/projects/{projectId}/storages/databases-analytics/data-streaming',
           },
           features: ['databases'],
           forceVisibility: true,
@@ -310,7 +311,8 @@ const pciNode: Node = {
           serviceType: 'CLOUD_PROJECT_DATABASE',
           routing: {
             application: 'public-cloud',
-            hash: '#/pci/projects/{projectId}/storages/databases-analytics/data-analysis',
+            hash:
+              '#/pci/projects/{projectId}/storages/databases-analytics/data-analysis',
           },
           features: ['databases'],
           forceVisibility: true,