+++ packages/manager/apps/key-management-service/src/data/api/orderCatalogOKMS.ts
@@ -1,0 +1,11 @@
+import apiClient from '@ovh-ux/manager-core-api';
+import { OKMSCatalog } from '@/types/orderCatalogOKMS.type';
+
+export const getOrderCatalogOKMS = async (
+  ovhSubsidiary: string,
+): Promise<OKMSCatalog> => {
+  const { data } = await apiClient.v6.get<OKMSCatalog>(
+    `/order/catalog/public/okms?ovhSubsidiary=${ovhSubsidiary}`,
+  );
+  return data;
+};