+++ packages/manager/apps/pci-rancher/src/pages/dashboard/edit-modal/EditModal.tsx
@@ -25,7 +25,8 @@ export default function EditModal() {
         onEditRancher={(r: RancherService) =>
           editRancherName({ rancher: r, editAction: EditAction.EditName })
         }
-        onClose={() => navigate('..')}      />
+        onClose={() => navigate('..')}
+      />
     </div>
   );
 }