+++ packages/manager/modules/iam/src/components/identities/identities.component.js
@@ -1,0 +1,13 @@
+import controller from './identities.controller';
+import template from './identities.template.html';
+
+export default {
+  bindings: {
+    identities: '<',
+    onRemove: '&',
+    onAdd: '&',
+    readOnly: '<',
+  },
+  controller,
+  template,
+};