+++ packages/manager/modules/models/BillingService/billing-service.constants.js
@@ -1,5 +1,8 @@
 export const DEBT_STATUS = ['PENDING_DEBT', 'UN_PAID', 'UNPAID'];
 
+export const BYOIP_SERVICE_PREFIX = 'byoip-failover-';
+
 export default {
   DEBT_STATUS,
+  BYOIP_SERVICE_PREFIX,
 };