+++ packages/manager/modules/octavia-load-balancer/src/load-balancers/load-balancer/logs/logs.component.js
@@ -1,0 +1,16 @@
+import template from './logs.template.html';
+import controller from './logs.controller';
+
+export default {
+  bindings: {
+    projectId: '<',
+    region: '<',
+    loadbalancerId: '<',
+    goToListingPage: '<',
+    logKindsList: '<',
+    kind: '<',
+    trackClick: '<',
+  },
+  template,
+  controller,
+};