+++ packages/manager/apps/pci-public-ip/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.12...@ovh-ux/manager-pci-public-ip-app@0.1.13) (2024-07-09)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-public-ip-app
+
+
+
+
+
+## [0.1.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.11...@ovh-ux/manager-pci-public-ip-app@0.1.12) (2024-07-03)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-public-ip-app
+
+
+
+
+
+## [0.1.11](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.10...@ovh-ux/manager-pci-public-ip-app@0.1.11) (2024-06-25)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-public-ip-app
+
+
+
+
+
+## [0.1.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.9...@ovh-ux/manager-pci-public-ip-app@0.1.10) (2024-06-20)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-public-ip-app
+
+
+
+
+
+## [0.1.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.8...@ovh-ux/manager-pci-public-ip-app@0.1.9) (2024-06-19)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-public-ip-app
+
+
+
+
+
 ## [0.1.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-public-ip-app@0.1.7...@ovh-ux/manager-pci-public-ip-app@0.1.8) (2024-06-17)