+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.5.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.5.0...@ovh-ux/manager-pci-app@4.5.1) (2024-07-09)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+# [4.5.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.4.12...@ovh-ux/manager-pci-app@4.5.0) (2024-07-04)
+
+
+### Features
+
+* add Logs to customer for Octavia Load Balancer ([#11860](https://github.com/ovh/manager/issues/11860)) ([0d66ac3](https://github.com/ovh/manager/commit/0d66ac3cbeccf4aa3c9464c08230077f1649c231))
+
+
+
+
+
+## [4.4.12](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.4.11...@ovh-ux/manager-pci-app@4.4.12) (2024-07-03)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.4.11](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.4.10...@ovh-ux/manager-pci-app@4.4.11) (2024-06-25)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [4.4.10](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.4.9...@ovh-ux/manager-pci-app@4.4.10) (2024-06-19)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [4.4.9](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@4.4.8...@ovh-ux/manager-pci-app@4.4.9) (2024-06-17)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app